requirements.txt
Flask==2.0.2
flask-cors==3.0.10
Pillow
numpy